pyproject.toml
[project]
name = "assessment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.2.2",
    "fastapi>=0.120.1",
    "openai-agents>=0.4.2",
    "pypdf>=6.1.3",
    "python-pptx>=1.0.2",
]
